feature/clockwise
I added a `clockwise` parameter for the direction of the text.

I simplified the code in the process: I had trouble understanding the concept of canvas rotation when not centered on the canvas center.
Now there's only an initial canvas translation to its center, and then multiple rotations (all based on the canvas center).